RST-6454 refactor testing project tree for better readability and str… (#133)
* RST-6454 refactor testing project tree for better readability and structure

* run uts and e2e separately in the build

* fix path to upload file helper

* fix circleci indentation

* run smoke test when browser is running in build

* separate health check from e2e

* remove warning from redirect test

* remove depreacted docker version from build

* fix import error for e2e tests

* fix e2e import error

* reroute reports files path

* add timeout to ping test to allow waitress to start up

* use args as fixtures

* remove warnings

* remove deprecation field flags warning

* remove unused validator
