+++ .circleci/config.yml
@@ -53,6 +53,7 @@ jobs:
           POSTGRES_PASSWORD: password
     steps:
       - checkout
+      - setup_remote_docker
       - when:
           condition:
             equal: [ DEV, << parameters.env_name >> ]
@@ -86,9 +87,6 @@ jobs:
                 name: Login to PROD ECR
                 command: |
                   aws ecr get-login-password --region $ECR_REGION_PROD | docker login --username AWS --password-stdin ${AWS_ECR_REGISTRY_ID_PROD}.dkr.ecr.${ECR_REGION_PROD}.amazonaws.com
-      - setup_remote_docker:
-          version: 20.10.11
-          docker_layer_caching: true
       - run:
           name: Install dependencies
           command: |
@@ -115,13 +113,19 @@ jobs:
             flask db migrate
             flask db upgrade
       - run:
-          name: Start site and test with pytest
+          name: Run unit and integration tests
+          command: |
+            . venv/bin/activate
+            pytest --ignore=tests/end_to_end_tests --ignore=tests/test_ping_json.py --html=reports/unit_int_report.html
+      - run:
+          name: Start site and run e2e
           command: |
             . venv/bin/activate
             waitress-serve --call --port=8080 'grc:create_app' &
-            pytest -s --show-capture=all -v --html=report.html
+            pytest tests/test_ping_json.py --max-retries 5 --timeout 3 --html=reports/smoke_report.html
+            pytest tests/end_to_end_tests/* -s --show-capture=all -v --html=reports/e2e_report.html
       - store_artifacts:
-          path: report.html
+          path: reports/
 
       - run:
           name: build app image
