+++ grc/utils/form_custom_validators.py
@@ -12,28 +12,6 @@ from grc.utils.reference_number import validate_reference_number
 from grc.models import db, Application
 
 
-class RequiredIf(DataRequired):
-    """Validator which makes a field required if another field is set and has a truthy value.
-
-    Sources:
-        - http://wtforms.simplecodes.com/docs/1.0.1/validators.html
-        - http://stackoverflow.com/questions/8463209/how-to-make-a-field-conditionally-optional-in-wtforms
-
-    """
-    field_flags = ('requiredif',)
-
-    def __init__(self, other_field_name, message=None, *args, **kwargs):
-        self.other_field_name = other_field_name
-        self.message = message
-
-    def __call__(self, form, field):
-        other_field = form[self.other_field_name]
-        if other_field is None:
-            raise Exception('no field named "%s" in form' % self.other_field_name)
-        if bool(other_field.data):
-            super(RequiredIf, self).__call__(form, field)
-
-
 class StrictRequiredIf(DataRequired):
     """Validator which makes a field required if another field is set and has a specific value.
 
@@ -42,7 +20,7 @@ class StrictRequiredIf(DataRequired):
         - http://stackoverflow.com/questions/8463209/how-to-make-a-field-conditionally-optional-in-wtforms
 
     """
-    field_flags = ('requiredif',)
+    field_flags = {'requiredif': True}
 
     def __init__(self, other_field_name, other_field_value, message=None, validators=None, *args, **kwargs):
         self.other_field_name = other_field_name
