+++ requirements.txt
@@ -15,7 +15,7 @@ email-validator==1.1.3
 Flask==2.3.3
 Flask-BabelEx==0.9.4
 Flask-Limiter[redis]==3.5.0
-Flask-Login==0.5.0
+Flask-Login==0.6.3
 Flask-Mail==0.9.1
 Flask-Migrate==3.1.0
 Flask-Principal==0.4.0
@@ -27,7 +27,7 @@ greenlet==1.1.2
 idna==3.3
 iniconfig==1.1.1
 itsdangerous==2.1.2
-Jinja2==3.1.2
+Jinja2==3.1.3
 jmespath==0.10.0
 jsonpickle~=2.1.0
 Mako==1.2.2
@@ -62,8 +62,8 @@ xhtml2pdf==0.2.15
 pillow==10.2.0
 PyMuPDF==1.20.1
 pyClamd==0.4.0
-cryptography==42.0.2
+cryptography==42.0.5
 aiozipstream==0.4
 smart-open==6.0.0
 pyAesCrypt==6.0.0
-sentry-sdk==1.14.0
+sentry-sdk==1.43.0
